add tests; refactor test organization
