--- src/utils/__tests__/fixtures/components5.json
@@ -1,27 +1,0 @@
-[
-  {
-    "label": "HTML",
-    "tag": "p",
-    "content": "",
-    "key": "html",
-    "type": "htmlelement"
-  },
-  {
-    "html": "<p>some text</p>",
-    "label": "Content",
-    "key": "content",
-    "type": "content"
-  },
-  {
-    "label": "Text Field",
-    "key": "textField",
-    "type": "textfield",
-    "input": true
-  },
-  {
-    "label": "Number",
-    "key": "number",
-    "type": "number",
-    "input": true
-  }
-]
